Add missing flannel placeholders to canal templates
The ConfigMap templates used for the flannel plugin include the ability
to specify the VXLAN VNI and port.

The equivalent templates for canal are missing the template fields for
these values so when specified in the cluster.yml, they are not
rendered.

This adds the missing template placeholders to the canal templates.