default-ingress-class and watch-ingress-without-class for ingress v1.1.0
Nginx ingress v1.1.0 template needs DefaultIngresClass, we added
this as a field in types for ingress options starting RKE v1.3.3.
But this field doesn't exist for RKE v1.2.x so using extraArgs
as a fallback way for users to update the field. This will allow
users to get the new template without upgrading Rancher via KDM.