[2.11] add the workflow for checking released versions (#1642)
This PR adds a GitHub workflow that checks whether any released RKE1, RKE2, and K3s versions are in the release-v2.x branches are missing in the local data.json file.
The workflow is triggered automatically at 2 AM, and can also be triggered manually in the GH UI.
A notification will be sent to the KDM channel in Slack if the workflow run fails.